download injectable

+++ flutter/lib/features/auth/domain/entities/medical_profile_entity.dart
@@ -1,7 +1,7 @@
 class MedicalProfileEntity {
   final String? bloodType;
-  final int? height;
-  final int? weight;
+  final double? height;
+  final double? weight;
   final List<String>? conditions;
   final List<String>? allergies;
   final List<RadiologyTestEntity>? radiologyTests;
